move test helper to crate::testing
